Apply cargo clippy --fix pass, resolve remaining correctness-level lints
Ran clippy --fix across the workspace with a broad feature set
(simulation, meshing, rendering, scripting-lua/js, ffi, serde, all store
backends) to catch lints gated behind optional code paths, then re-ran
until no more auto-fixes were offered. Covers idiom cleanups like
is_multiple_of() over manual %, mem::take over mem::replace(_, T::new()),
assert! over assert_eq!(_, bool), and redundant struct-update fields.

Also fixed the small number of deny-by-default correctness lints that
clippy --fix can't apply automatically:
- region.rs: dropped always-true `index >= 0` checks on a usize
- formats/anvil.rs: a chunk-index test asserted trivial tautologies
  (0*32==0, 32==32) instead of exercising the formula
- approx_constant false positives on arbitrary test float literals
  (3.14, 3.14159) that aren't meant to represent pi — allowed where
  the value is genuinely arbitrary, switched to std::f32::consts::PI
  where a test comment specifically says "test pi"

Fixed an unrelated pre-existing build break surfaced while compiling
under the `rendering` feature: examples/render_schematic.rs's
CameraConfig literals predated the target/projection/background fields
added in v0.2.11 and never got updated, since no CI job builds examples
with --features rendering.

Left the ~735 "might dereference a raw pointer but not marked unsafe"
lints in ffi.rs alone — marking every pub extern "C" fn unsafe is a
large, mechanical change worth doing as its own pass alongside the
ffi.rs module split (tracked separately), not folded into a lint
cleanup commit.

+++ examples/test_unicode_circuit.rs
@@ -57,9 +57,9 @@ fn main() -> Result<(), String> {
     // Print some sample blocks to verify
     println!("\n🔍 Sample blocks (first 15 non-air):");
     let mut count = 0;
-    'outer: for y in 0..height as i32 {
-        for z in 0..depth as i32 {
-            for x in 0..width as i32 {
+    'outer: for y in 0..height {
+        for z in 0..depth {
+            for x in 0..width {
                 if let Some(block) = half_adder.get_block(x, y, z) {
                     let block_str = block.to_string();
                     if !block_str.contains("air") {
@@ -113,8 +113,8 @@ fn main() -> Result<(), String> {
 
     // Show what blocks were created
     println!("\n   Blocks in repeater chain:");
-    for y in 0..h as i32 {
-        for x in 0..w as i32 {
+    for y in 0..h {
+        for x in 0..w {
             if let Some(block) = repeater_chain.get_block(x, y, 0) {
                 let block_str = block.to_string();
                 if !block_str.contains("air") {
